requirements.txt
praw==7.7.1
pandas==2.2.1
python-dotenv==1.0.0
openai==1.12.0
numpy==1.26.4 